# swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: API Information
servers:
  - url: http://localhost:4000
paths:
  /signup:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created successfully
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  token:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /books:
    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                book:
                  type: string
                  format: binary
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Book created successfully
        '400':
          description: Bad request
        '500':
          description: Internal server error
    get:
      summary: Get all books
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          description: Bad request
  /books/{id}:
    get:
      summary: Get a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single book
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Book not found
        '500':
          description: Internal server error
    put:
      summary: Update a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                book:
                  type: string
                  format: binary
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Book updated successfully
        '400':
          description: Bad request
        '403':
          description: Unauthorized request
        '404':
          description: Book not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book deleted successfully
        '401':
          description: Not authorized
        '500':
          description: Internal server error
  /books/best-rated:
    get:
      summary: Get best rated books
      responses:
        '200':
          description: A list of best rated books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '500':
          description: Internal server error
  /books/{id}/rating:
    post:
      summary: Add a rating to a book
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                rating:
                  type: integer
      responses:
        '200':
          description: Rating added successfully
        '400':
          description: Bad request
        '404':
          description: Book not found
        '500':
          description: Internal server error